+++ jsonschema/tests/test_jsonschema_test_suite.py
@@ -28,14 +28,14 @@ def skip(message, **kwargs):
     return skipper
 
 
-def missing_format(checker):
+def missing_format(Validator):
     def missing_format(test):  # pragma: no cover
         schema = test.schema
         if (
             schema is True
             or schema is False
             or "format" not in schema
-            or schema["format"] in checker.checkers
+            or schema["format"] in Validator.FORMAT_CHECKER.checkers
             or test.valid
         ):
             return
@@ -150,10 +150,10 @@ TestDraft3 = DRAFT3.to_unittest_testcase(
     DRAFT3.optional_tests_of(name="non-bmp-regex"),
     DRAFT3.optional_tests_of(name="zeroTerminatedFloats"),
     Validator=jsonschema.Draft3Validator,
-    format_checker=jsonschema.draft3_format_checker,
+    format_checker=jsonschema.Draft3Validator.FORMAT_CHECKER,
     skip=lambda test: (
         narrow_unicode_build(test)
-        or missing_format(jsonschema.draft3_format_checker)(test)
+        or missing_format(jsonschema.Draft3Validator)(test)
         or complex_email_validation(test)
         or skip(
             message=bug(),
@@ -175,12 +175,12 @@ TestDraft4 = DRAFT4.to_unittest_testcase(
     DRAFT4.optional_tests_of(name="non-bmp-regex"),
     DRAFT4.optional_tests_of(name="zeroTerminatedFloats"),
     Validator=jsonschema.Draft4Validator,
-    format_checker=jsonschema.draft4_format_checker,
+    format_checker=jsonschema.Draft4Validator.FORMAT_CHECKER,
     skip=lambda test: (
         narrow_unicode_build(test)
         or allowed_leading_zeros(test)
         or leap_second(test)
-        or missing_format(jsonschema.draft4_format_checker)(test)
+        or missing_format(jsonschema.Draft4Validator)(test)
         or complex_email_validation(test)
         or skip(
             message=bug(),
@@ -236,12 +236,12 @@ TestDraft6 = DRAFT6.to_unittest_testcase(
     DRAFT6.optional_tests_of(name="float-overflow"),
     DRAFT6.optional_tests_of(name="non-bmp-regex"),
     Validator=jsonschema.Draft6Validator,
-    format_checker=jsonschema.draft6_format_checker,
+    format_checker=jsonschema.Draft6Validator.FORMAT_CHECKER,
     skip=lambda test: (
         narrow_unicode_build(test)
         or allowed_leading_zeros(test)
         or leap_second(test)
-        or missing_format(jsonschema.draft6_format_checker)(test)
+        or missing_format(jsonschema.Draft6Validator)(test)
         or complex_email_validation(test)
         or skip(
             message=bug(),
@@ -260,12 +260,12 @@ TestDraft7 = DRAFT7.to_unittest_testcase(
     DRAFT7.optional_tests_of(name="float-overflow"),
     DRAFT7.optional_tests_of(name="non-bmp-regex"),
     Validator=jsonschema.Draft7Validator,
-    format_checker=jsonschema.draft7_format_checker,
+    format_checker=jsonschema.Draft7Validator.FORMAT_CHECKER,
     skip=lambda test: (
         narrow_unicode_build(test)
         or allowed_leading_zeros(test)
         or leap_second(test)
-        or missing_format(jsonschema.draft7_format_checker)(test)
+        or missing_format(jsonschema.Draft7Validator)(test)
         or complex_email_validation(test)
         or skip(
             message=bug(),
@@ -408,12 +408,12 @@ TestDraft201909 = DRAFT201909.to_unittest_testcase(
 TestDraft201909Format = DRAFT201909.to_unittest_testcase(
     DRAFT201909.format_tests(),
     Validator=jsonschema.Draft201909Validator,
-    format_checker=jsonschema.draft201909_format_checker,
+    format_checker=jsonschema.Draft201909Validator.FORMAT_CHECKER,
     skip=lambda test: (
         complex_email_validation(test)
         or allowed_leading_zeros(test)
         or leap_second(test)
-        or missing_format(jsonschema.draft201909_format_checker)(test)
+        or missing_format(jsonschema.Draft201909Validator)(test)
         or complex_email_validation(test)
     ),
 )
@@ -533,12 +533,12 @@ TestDraft202012 = DRAFT202012.to_unittest_testcase(
 TestDraft202012Format = DRAFT202012.to_unittest_testcase(
     DRAFT202012.format_tests(),
     Validator=jsonschema.Draft202012Validator,
-    format_checker=jsonschema.draft202012_format_checker,
+    format_checker=jsonschema.Draft202012Validator.FORMAT_CHECKER,
     skip=lambda test: (
         complex_email_validation(test)
         or allowed_leading_zeros(test)
         or leap_second(test)
-        or missing_format(jsonschema.draft202012_format_checker)(test)
+        or missing_format(jsonschema.Draft202012Validator)(test)
         or complex_email_validation(test)
     ),
 )